server/pyproject.toml
[tool.poetry]
name = "fse_trab_1_server"
version = "0.1.0"
description = ""
authors = ["Eduardo-RFarias <eduardo.rfarias@outlook.com>"]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.10"
python-socketio = "^5.11.2"
aiohttp = {extras = ["speedups"], version = "^3.9.5"}
